feat(frontend): pass onkeydown to the input tm-388
